Implement two helper-functions for casting spell in creatureAI.
Function will return a cast result, with fail reasons if spell can not be casted (can be used to take alternative action). Please note this is not a full list of cast fail errors/reasons, but will cover a few basic errors that may occur.
The virtual function CanCastSpell() may be adjusted for custom needs from script library if not the default is sufficient.

Signed-off-by: NoFantasy <nofantasy@nf.no>

(based on master commit 822255c)

(based on commit 9847816)

Signed-off-by: VladimirMangos <vladimir@getmangos.com>