[z1634] Implement CONDITION_LEARNABLE_ABILITY, check by spell and optionally item
The primary use are intended for items that teach spell, typically for primary/secondary tradeskills.

Condition is true when player does not know the spell already, has the needed corresponding skill (and skill level) and does not have the item already (when item is defined).

Signed-off-by: NoFantasy <nofantasy@nf.no>

(based on master commit c9c6ec5)

(based on commit d6d9294)